my-server/pyproject.toml
[project]
name = "my-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Punyaphat Kaewkao",email = "punyaphat0010@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic[email] (>=2.11.9,<3.0.0)",
    "fastapi (>=0.117.1,<0.118.0)",
    "uvicorn (>=0.36.0,<0.37.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "sqlmodel (>=0.0.25,<0.0.26)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]

[tool.poetry]
packages = [{include = "my_server", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
